switched to installing the Helm chart directly
Rather than install using a HelmChart object type and the Helm operator to install the operator now the NVIDIA GPU Operator Helm repo is added and the Helm chart is installed directly from the repo. The function that installed using a HelmChart object type is still included in the code.