enable configuration of a cluster virtual IP
The RKE2 deployment scripts now have the ability to also install kube-vip which allows for the configuration of a cluster virtual IP (VIP). This has also necessitated updating the .cfg files with additional variables and updating some of the other service deployment scripts to explicitly specify the ingress class to use when creating their ingress. The cluster VIP has been tested and validated with Rancher but has not yet been tested with the other services. The default is to _not_ enabled the cluster VIP so things should still work the same as before.